requirements.txt
requests
beautifulsoup4
bs4
google-genai
